Opening files for the interface within the client of the 3 interfaces and changing the name of the authProvider file
