+++ clinet/src/interface/Aadress.tsx
@@ -1,0 +1,5 @@
+export interface IAddress extends Document {
+  city: string;
+  street: string;
+  buildingNumber: number;
+}
